+++ psutil/_psosx.py
@@ -60,9 +60,6 @@ svmem = namedtuple(
 
 pmem = namedtuple('pmem', ['rss', 'vms', 'pfaults', 'pageins'])
 pfullmem = namedtuple('pfullmem', pmem._fields + ('uss', ))
-scpustats = namedtuple(
-    'scpustats', ['ctx_switches', 'interrupts', 'soft_interrupts',
-                  'syscalls', 'traps'])
 
 pmmap_grouped = namedtuple(
     'pmmap_grouped',
@@ -126,8 +123,8 @@ def cpu_count_physical():
 def cpu_stats():
     ctx_switches, interrupts, soft_interrupts, syscalls, traps = \
         cext.cpu_stats()
-    return scpustats(
-        ctx_switches, interrupts, soft_interrupts, syscalls, traps)
+    return _common.scpustats(
+        ctx_switches, interrupts, soft_interrupts, syscalls)
 
 
 def boot_time():