+++ psutil/_pswindows.py
@@ -104,8 +104,6 @@ ntpinfo = namedtuple(
     'ntpinfo', ['num_handles', 'ctx_switches', 'user_time', 'kernel_time',
                 'create_time', 'num_threads', 'io_rcount', 'io_wcount',
                 'io_rbytes', 'io_wbytes'])
-ssysinfo = namedtuple(
-    'ssysinfo', ['ctx_switches', 'interrupts', 'dpcs', 'syscalls'])
 
 
 # set later from __init__.py
@@ -221,7 +219,9 @@ def cpu_count_physical():
 def cpu_stats():
     """Return CPU statistics."""
     ctx_switches, interrupts, dpcs, syscalls = cext.cpu_stats()
-    return ssysinfo(ctx_switches, interrupts, dpcs, syscalls)
+    soft_interrupts = 0
+    return _common.scpustats(ctx_switches, interrupts, soft_interrupts,
+                             syscalls)
 
 
 def boot_time():