+++ psutil/tests/test_bsd.py
@@ -297,9 +297,9 @@ class FreeBSDSpecificTestCase(unittest.TestCase):
         self.assertAlmostEqual(psutil.cpu_stats().syscalls,
                                sysctl('vm.stats.sys.v_syscall'), delta=1000)
 
-    def test_cpu_stats_traps(self):
-        self.assertAlmostEqual(psutil.cpu_stats().traps,
-                               sysctl('vm.stats.sys.v_trap'), delta=1000)
+    # def test_cpu_stats_traps(self):
+    #    self.assertAlmostEqual(psutil.cpu_stats().traps,
+    #                           sysctl('vm.stats.sys.v_trap'), delta=1000)
 
 
 # =====================================================================